+++ test/unit/fuzz-tools.test.js
@@ -311,6 +311,7 @@ import * as fitzpatrickV331 from '../../lib/fitzpatrick-v331.js';
 import * as haggittV332 from '../../lib/haggitt-v332.js';
 import * as kikuchiV333 from '../../lib/kikuchi-v333.js';
 import * as kudoV334 from '../../lib/kudo-v334.js';
+import * as niceV335 from '../../lib/nice-v335.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -567,6 +568,7 @@ const MODULES = {
   'haggitt-v332.js': haggittV332,
   'kikuchi-v333.js': kikuchiV333,
   'kudo-v334.js': kudoV334,
+  'nice-v335.js': niceV335,
   'tb-testing.js': tbTesting,
 };